drivers: sensor: dht: Fix sensor read error on DHT11/22 on esp_devkitc_wrover
It was observed during local testing on an esp32 (devkitc-wrover) that
explicitly setting the logical level of the configured GPIO to false
causes timing issues that lead to "dht_measure_signal_duration" missing
signals during the window after the sensor is asserted for 18ms. This
fix in intended to avoid consistent read failures that make the driver
fail on ESP32.

Signed-off-by: Omar Naffaa <omarnaffaa.on@gmail.com>